Add controls for GXNext and future GeneXus' releases (#48)
* Add Component interface

* Add drag bar control

This control defines a simple container with two slots that are separated by a draggable vertical bar, which can resize both slots.

At this moment, it only support slots separation via a vertical bar. In the future, it might support a horizontal bar to resize the height.

* Add RTL support for drag-bar control

* Add ch-next-progress-bar

A determinate progress bar that measures progress using "steps".

The progress bar has a title (caption property) and a description. When the description changes at runtime, the state transition is animated with a fade-out and fade-in.