fix ch-window hide logic
Sometimes ch-window was not being displayed again, when it should. The issue was that the display of the ch-window is bound to whether or not there is slotted content, and there were a couple of situations where slotted content remained when it shouldn't:
- when the ch-window was closed by the user, by pressing escape, or by clicking outside the window
- when the ch-window was closed by the user because he/she selected an item.

The fix is to clear slotted content every time the ch-window is closed.