add 'selectionChanged' event
Even though the suggest-list-item emmits an event 'itemSelected' that can be used by the userm it is not possible to use this event directly on the ch-suggest element like this:

<ch-suggest onItemSelected="..." >

because the event does not exists on the ch-suggest. To improve how the user can listen to the event, an additional event was created on the ch-suggest, called 'selectionChanged', that emits the same information the 'itemSelected' emits.