fix ch-window hide logic (#160)
* fix ch-window hide logic

Sometimes ch-window was not being displayed again, when it should. The issue was that the display of the ch-window is bound to whether or not there is slotted content, and there were a couple of situations where slotted content remained when it shouldn't:
- when the ch-window was closed by the user, by pressing escape, or by clicking outside the window
- when the ch-window was closed by the user because he/she selected an item.

The fix is to clear slotted content every time the ch-window is closed.

* add 'selectionChanged' event

Even though the suggest-list-item emmits an event 'itemSelected' that can be used by the userm it is not possible to use this event directly on the ch-suggest element like this:

<ch-suggest onItemSelected="..." >

because the event does not exists on the ch-suggest. To improve how the user can listen to the event, an additional event was created on the ch-suggest, called 'selectionChanged', that emits the same information the 'itemSelected' emits.

* prevent itemSelected event propagation

- prevent itemSelected event propagation
- emit event.detail directly for the selectionChanged event

* test propagation prevention