Refactor the combo-box implementation
Fixed broken style when using the control on mobile devices.

Fixed broken support for suggestOptions = { alreadyProcessed: true }

Fixed an issue where the "disabled" part was not propertly displayed in the "group__header" part.

Implemented event delegation to improve the rendering performance.

Added name property to improve form support.

Fix: Revert changes when the filters are strict and the popover is closed with Escape key or external click.

Breaking changes:
 - Removed the filter property and filterChange event. Now, when suggest = true, changes to the input value will update the value property and emit the input event.

 - Now the caption property of the items can be undefined. If undefined, the value will be used as a replacement of the caption when rendering the item.