Fix selection="single" issues
Fixed selection="single" not working when the model is set with a selected item and the user selects a different item.

Fixed selection="single" not working when the model is set with a selected item and the updateItemProperties method is called to select a different item.

Fixed selection="single" not working when the model is updated at runtime.