Add a utility to report accessibility issues in form components and fix accessibility issues in the `ch-checkbox`, `ch-slider` and `ch-switch` components (#486)
* Use the visible label instead of the accessibleName in the ch-checkbox, ch-slider and ch-switch components

When both the visible label and accessibleName property are set, the visible label will take precedence over the accessibleName property since it is a visible label.

* Add a utility to report accessibility issues in form components

This utility check if form elements have a valid label and report when they don't acomplish this condition. At this moment, it's supported in the following component:
 - ch-checkbox
 - ch-combo-box-render
 - ch-edit
 - ch-slider
 - ch-switch

* Create analyzeLabelExistence.e2e.ts

* Update unit tests

* Add more unit tests

* Disable the accessibility reports in the ch-checkbox and ch-switch components

We need to investigate these components further to implement accessibility reports

* Add support to disable accessibility reports

* Update readmes

* [showcase] Disable accessibility checker in prod builds