`[ch-dialog]` Add support for disabling close actions (#533)
* `[ch-dialog]` Add support for disabling close actions

Added the `closable` property in the `ch-dialog`.

If `closable === true` the `ch-dialog` can be closed by user interaction, in which case a close button is rendered in the header (when `showHeader === true`).

If `closable === false`, the close button is not rendered and pressing the `Esc` key or clicking outside of the `ch-dialog` will not close it.

* Update readme.md

* Update showcase story