`[ch-tab-render]` Add support to apply ellipsis when the tab caption overflows the max-inline-size
Now, tabs of the `ch-tab-render` component automatically apply an ellipsis when the caption doesn't fit the container size. This case can occur when applying a max-inline-size for the `tab` part.

When the text is displayed with an ellipsis, hovering the tab caption will display a tooltip with the full text.