Mark "lit" dependency as external to optimize bundle size
The "lit" dependency is now marked as external, meaning it won't be bundled with the library. This improves code reuse if another library that uses chameleon also uses Lit, and can also fix some issues as it won't mix Lit versions for the development server (because Chameleon was bundling Lit in prod mode, but another libraries use Lit in "dev" mode for the dev server).